add safe mode guards

+++ main.py
@@ -7,6 +7,7 @@ from datetime import datetime, timedelta, timezone
 from dotenv import load_dotenv
 import os
 
+
 # Either load the env vars from a local env file (when testing stuff locally) or it's already been loaded (running on server)
 if os.path.isfile('.env'):
     load_dotenv()
@@ -50,12 +51,16 @@ async def deleteHere(ctx):
         await channel.send('can\'t recognise that time format, please enter it in days, hours, or minutes with for example `5d`, `12h`, or `10m`')
         return
 
+    if SAFEMODE:
+        await channel.send(f'SAFEMODE: would have added this channel for future auto purging with max message age of {getReadableMaxAge(maxAge)}')
+        return
+
     # Attempt to add channel to database
     result = database.addChannel(channel.id, maxAge)
     if result == 1:
         await channel.send('sorry, something went wrong :/ please try again (it will probably work this time :3 )')
     else:
-        await channel.send('oki! (the first purge at this max age will be in the next first purge cycle :3 )')
+        await channel.send(f'oki! Auto purging set up with max message age of {getReadableMaxAge(maxAge)} (the first purge in this channel will be in the next first purge cycle :3 )')
 
 
 def parseMaxAge(message: str) -> int:
@@ -107,6 +112,9 @@ async def delete():
     for (channelID, maxAge) in database.getChannels():
         cutoff = now - timedelta(minutes=maxAge)
         channel = client.get_channel(channelID)
+        if SAFEMODE:
+            print(f'SAFEMODE: Would purge all messages in channel {channelID} before {cutoff} (UTC)')
+            continue
         
         print(f'Purging all messages in channel {channelID} before {cutoff} (UTC)')
         await channel.purge(limit=None, before=cutoff, reason='included in periodic channel purge', oldest_first=True)
@@ -117,6 +125,10 @@ async def delete():
 async def stopDeletion(ctx):
     """Command to deactivate auto purging in a channel"""
     channel = ctx.channel
+    if SAFEMODE:
+        await channel.send('SAFEMODE: would have removed this channel from auto purges')
+        return
+
     result = database.deleteChannel(channel.id)
     if result == 1:
         await channel.send('sorry, something went wrong :/ please try again (it will probably work this time :3 )')
@@ -143,15 +155,20 @@ async def info(ctx):
     # Find the channel if possible
     for (channelID, maxAge) in channels:
         if channelID == channel.id:
-            if (maxAge/24)/60 == (maxAge//24)//60:
-                await channel.send(f'This channel auto deletes messages that are more than {maxAge//24//60} days old')
-            elif maxAge/60 == maxAge//60:
-                await channel.send(f'This channel auto deletes messages that are more than {maxAge//60} hours old')
-            else:
-                await channel.send(f'This channel auto deletes messages that are more than {maxAge} minutes old')
+            await channel.send(f'This channel auto deletes messages that are more than {getReadableMaxAge(maxAge)} old')
             return
     
     await channel.send(f'This channel does not have auto deletion set up')
+
+
+def getReadableMaxAge(minutes: int) -> str:
+    """Returns a human readable conversion of minutes to days or hours or minutes"""
+    if minutes/24/60 == minutes//24//60:
+        return f'{minutes//24//60} days'
+    elif minutes/60 == minutes//60:
+        return f'{minutes//60} hours'
+    else:
+        return f'{minutes} minutes'
 
 
 TOKEN = os.environ.get('TOKEN')
